ci: split long-running unit tests These changes only reorganize existing unit tests into separate CI jobs. No production code or test logic changed.
